moving functions also used by AdvPatch to a utils file
